Fix WRONG_OBJ: ClpNode mini-BAB stored integer bounds overflow as int
Root-caused a wrong-optimal regression on MIPLIB2017 gen-ip054 (Cbc claimed
optimal at ~6848-6853 instead of the certified 6840.96564179) to
ClpNode::lower_/upper_ storing general-integer variable bounds as 32-bit
int in gutsOfConstructor()/applyNode() (used by fathomMany()'s mini-BAB
node storage/restore, active whenever CbcModel's fastNodeDepth_ path is
engaged - i.e. whenever -threads is not explicitly set, or set to 0).

gen-ip054 has general-integer variables with no explicit MPS upper bound,
which Cbc defaults to a large sentinel (~1.23456789e10). That value
overflows a 32-bit int (max ~2.1e9), silently wrapping to a garbage
negative bound (observed: -2147483648) when a mini-BAB node's bounds were
cast via static_cast<int>(...) for storage and later read back via
applyNode(..., 3). The corrupted bound permanently excludes the subtree
containing the true optimum from ever being (re-)explored, so Cbc
converges on and proves a worse objective as 'optimal'.

Fix: widen ClpNode::lower_/upper_ from int* to double* (doubles losslessly
represent integer values up to 2^53, far beyond any bound Cbc's big-M
sentinel or any realistic instance would use), removing the
static_cast<int> truncation at both the storage and restore sites.

Verified via mip-sanity-data's gen-ip054 fixture and direct CLI repro
(gen-ip054, -sec 180 -maxNodes 2000000, no -threads flag, matching the
CLI's default numberBeforeTrust_=10): before this fix, falsely reports
proven-optimal 6852-6853 depending on numberBeforeTrust_; after the fix,
correctly reports 'Stopped (node limit)' with a legitimate ~1% gap
instead of a false proof of optimality. Full mip-sanity-data suite (473
checks) passes with zero regressions.

+++ src/ClpNode.cpp
@@ -447,14 +447,14 @@ void ClpNode::gutsOfConstructor(ClpSimplex *model, const ClpNodeStuff *stuff,
       delete[] lower_;
       delete[] upper_;
       maximumIntegers_ = iInteger;
-      lower_ = new int[maximumIntegers_];
-      upper_ = new int[maximumIntegers_];
+      lower_ = new double[maximumIntegers_];
+      upper_ = new double[maximumIntegers_];
     }
     iInteger = 0;
     for (iColumn = 0; iColumn < numberColumns; iColumn++) {
       if (integerType[iColumn]) {
-        lower_[iInteger] = static_cast< int >(lower[iColumn]);
-        upper_[iInteger] = static_cast< int >(upper[iColumn]);
+        lower_[iInteger] = lower[iColumn];
+        upper_[iInteger] = upper[iColumn];
         iInteger++;
       }
     }
@@ -536,8 +536,8 @@ void ClpNode::createArrays(ClpSimplex *model)
     delete[] lower_;
     delete[] upper_;
     maximumIntegers_ = numberIntegers;
-    lower_ = new int[numberIntegers];
-    upper_ = new int[numberIntegers];
+    lower_ = new double[numberIntegers];
+    upper_ = new double[numberIntegers];
   }
 }
 // Clean up as crunch is different model
@@ -590,9 +590,9 @@ void ClpNode::applyNode(ClpSimplex *model, int doBoundsEtc)
     for (int iColumn = 0; iColumn < numberColumns; iColumn++) {
       if (integerType[iColumn]) {
         iInteger++;
-        if (lower_[iInteger] != static_cast< int >(lower[iColumn]))
+        if (lower_[iInteger] != lower[iColumn])
           model->setColumnLower(iColumn, lower_[iInteger]);
-        if (upper_[iInteger] != static_cast< int >(upper[iColumn]))
+        if (upper_[iInteger] != upper[iColumn])
           model->setColumnUpper(iColumn, upper_[iInteger]);
       }
     }
